+++ components/frontend/src/widgets/TableRowWithDetails.jsx
@@ -1,11 +1,20 @@
-import { TableCell, TableRow } from "@mui/material"
+import { Stack, TableCell, TableRow } from "@mui/material"
 import { bool, func, string } from "prop-types"
+import React from "react"
 
 import { childrenPropType } from "../sharedPropTypes"
 import { ExpandButton } from "./buttons/ExpandButton"
 
 export function TableRowWithDetails(props) {
     const { color, children, details, expanded, onExpand, ...otherProps } = props
+    const [firstTableCell, ...otherTableCells] = React.Children.toArray(children)
+    const firstTableCellWithExpandButton = React.cloneElement(
+        firstTableCell,
+        null,
+        <Stack direction="row" alignItems="center">
+            <ExpandButton expand={expanded} onClick={() => onExpand(!expanded)} /> {firstTableCell.props.children}
+        </Stack>,
+    )
     return (
         <>
             <TableRow
@@ -20,10 +29,8 @@ export function TableRowWithDetails(props) {
                     },
                 }}
             >
-                <TableCell align="center" padding="checkbox" /* Make sure the column does not stretch */>
-                    <ExpandButton expand={expanded} onClick={() => onExpand(!expanded)} />
-                </TableCell>
-                {children}
+                {firstTableCellWithExpandButton}
+                {otherTableCells}
             </TableRow>
             {expanded && (
                 <TableRow>