+++ components/frontend/src/widgets/buttons/ExpandButton.jsx
@@ -1,16 +1,26 @@
 import { IconButton } from "@mui/material"
-import { bool, func } from "prop-types"
+import { bool, func, string } from "prop-types"
 
 import { CaretDown, CaretRight } from "../icons"
 
-export function ExpandButton({ expand, onClick }) {
+export function ExpandButton({ expand, onClick, size }) {
     return (
-        <IconButton aria-label="Expand/collapse" onClick={onClick}>
-            {expand ? <CaretDown size="1.5em" /> : <CaretRight size="1.5em" />}
+        <IconButton
+            aria-label="Expand/collapse"
+            onClick={onClick}
+            sx={{
+                "@media print": {
+                    display: "none !important",
+                },
+                marginLeft: "-0.5rem",
+            }}
+        >
+            {expand ? <CaretDown size={size ?? "1.5em"} /> : <CaretRight size={size ?? "1.5em"} />}
         </IconButton>
     )
 }
 ExpandButton.propTypes = {
     expand: bool,
     onClick: func,
+    size: string,
 }